+++ src/tests/Shape.test.mjs
@@ -1,21 +1,21 @@
-import { Data, isData, typeName } from "../Data.mjs";
-import { Trait } from "../Trait.mjs";
+import { Data, isData, variantName, Trait } from "../index.mjs";
 
 describe('Shape tests', () => {
-    const Shape = Data({ Circle: ['radius'], Rectangle: ['width', 'height'] });
+    const Shape = Data({ Circle: ['radius'], Rectangle: ['width', 'height'] }),
+        { Circle, Rectangle } = Shape;
 
     test('Shape Data', () => {
         expect(Shape[isData]).toBe(true);
 
-        const circle = Shape.Circle({ radius: 1 });
-        expect(circle[typeName]).toBe('Circle');
+        const circle = Circle({ radius: 1 });
+        expect(circle[variantName]).toBe('Circle');
         expect(circle).toBeDefined();
         expect(circle.radius).toBe(1);
         expect(circle.width).toBeUndefined();
         expect(circle.height).toBeUndefined();
 
-        const rectangle = Shape.Rectangle({ width: 2, height: 3 });
-        expect(rectangle[typeName]).toBe('Rectangle');
+        const rectangle = Rectangle({ width: 2, height: 3 });
+        expect(rectangle[variantName]).toBe('Rectangle');
         expect(rectangle).toBeDefined();
         expect(rectangle.radius).toBeUndefined();
         expect(rectangle.width).toBe(2);
@@ -28,10 +28,10 @@ describe('Shape tests', () => {
     })
 
     test('Shape area', () => {
-        const circle = Shape.Circle({ radius: 1 });
+        const circle = Circle({ radius: 1 });
         expect(area(circle)).toBe(Math.PI);
 
-        const rectangle = Shape.Rectangle({ width: 2, height: 3 });
+        const rectangle = Rectangle({ width: 2, height: 3 });
         expect(area(rectangle)).toBe(6);
     })
 })